+++ src/main/java/com/project/sooktoring/user/profile/controller/UserProfileController.java
@@ -3,6 +3,7 @@ package com.project.sooktoring.user.profile.controller;
 import com.project.sooktoring.user.auth.util.CurrentUser;
 import com.project.sooktoring.user.auth.util.UserPrincipal;
 import com.project.sooktoring.user.profile.dto.request.UserProfileRequest;
+import com.project.sooktoring.user.profile.dto.response.MentorResponse;
 import com.project.sooktoring.user.profile.dto.response.UserProfileResponse;
 import com.project.sooktoring.user.profile.service.UserProfileService;
 import io.swagger.annotations.Api;
@@ -17,7 +18,7 @@ import java.util.List;
 import static org.springframework.http.MediaType.*;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/user/profiles")
 @RequiredArgsConstructor
 @Api(tags = "유저 프로필 API")
 public class UserProfileController {
@@ -25,7 +26,7 @@ public class UserProfileController {
     private final UserProfileService userProfileService;
 
     @Operation(summary = "나의 프로필 상세조회", description = "나의 프로필 상세조회")
-    @GetMapping("/profile")
+    @GetMapping("/me")
     public UserProfileResponse getMyUserProfile(@CurrentUser UserPrincipal currentUser) {
         return userProfileService.getUserProfile(currentUser.getUserId());
     }
@@ -33,7 +34,7 @@ public class UserProfileController {
     //프로필 등록 API 사용X -> 회원가입 시 자동으로 기본값으로 초기화 -> 이후 계속 수정만 가능
 
     @Operation(summary = "나의 프로필 수정", description = "나의 프로필 수정")
-    @PutMapping(value = "/profile", consumes = {APPLICATION_JSON_VALUE, MULTIPART_FORM_DATA_VALUE})
+    @PutMapping(value = "/me", consumes = {APPLICATION_JSON_VALUE, MULTIPART_FORM_DATA_VALUE})
     public String updateMyUserProfile(@RequestPart @Validated UserProfileRequest userProfileRequest,
                                       @RequestPart MultipartFile file,
                                       @CurrentUser UserPrincipal currentUser) {
@@ -42,16 +43,26 @@ public class UserProfileController {
     }
 
     @Operation(summary = "모든 이용자 프로필 조회", description = "모든 이용자 프로필 조회")
-    @GetMapping("/profiles")
+    @GetMapping
     public List<UserProfileResponse> getUserProfiles() {
         return userProfileService.getUserProfiles();
     }
 
     @Operation(summary = "특정 이용자 프로필 상세조회", description = "아이디 유효하지 않으면 null 반환")
-    @GetMapping("/profiles/{userId}")
+    @GetMapping("/{userId}")
     public UserProfileResponse getUserProfile(@PathVariable Long userId) {
         return userProfileService.getUserProfile(userId);
     }
 
+    @Operation(summary = "멘토들의 프로필 조회", description = "멘토 아이디, 이름, 직업, 연차, 주전공")
+    @GetMapping("/mentors")
+    public List<MentorResponse> getMentorList() {
+        return userProfileService.getMentorList();
+    }
 
+    @Operation(summary = "특정 멘토의 프로필 조회", description = "멘토 아이디, 이름, 직업, 연차, 주전공")
+    @GetMapping("/mentors/{mentorId}")
+    public MentorResponse getMentor(@PathVariable Long mentorId) {
+        return userProfileService.getMentor(mentorId);
+    }
 }
